API Endpoint Connection
Connected Registration and Notification APIs to deployed Backend, Minor UI Improvements, Login Page debug
