+++ android/app/src/google-services.json
@@ -29,6 +29,13 @@
             {
               "client_id": "319149493944-lmosinhbjme3o02rodscjhe4rhan2o4e.apps.googleusercontent.com",
               "client_type": 3
+            },
+            {
+              "client_id": "319149493944-62mdf8kkt2bj52fhedgoi3oetl66jeca.apps.googleusercontent.com",
+              "client_type": 2,
+              "ios_info": {
+                "bundle_id": "com.example.dashboardapp"
+              }
             }
           ]
         }
