+++ lib/resourcespage.dart
@@ -6,7 +6,7 @@ class ResourcesPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Center(
-      child: Text('Page 2 Content'),
+      child: Text('Coming Soon!'),
     );
   }
 }
