+++ lib/socialsscreen.dart
@@ -6,7 +6,7 @@ class SocialsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Center(
-      child: Text('Page 4 Content'),
+      child: Text('Coming Soon!'),
     );
   }
 }
