--- lib/userhome.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class UserHome extends StatelessWidget {
-  const UserHome({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Center(
-      child: Text('Page 1 Content'),
-    );
-  }
-}
